# Conflicts:
#	crypto/signing/kyber/multisig/bls_test.go
#	go.sum
#	process/peer/process.go

+++ node/heartbeat/monitorEdgeCases_test.go
@@ -473,6 +473,6 @@ func verifyHeartBeat(
 	downTimeSeconds int,
 ) {
 	assert.Equal(t, isActive, heartBeat.IsActive)
-	assert.Equal(t, upTimeSeconds, heartBeat.TotalUpTime)
-	assert.Equal(t, downTimeSeconds, heartBeat.TotalDownTime)
+	assert.Equal(t, int64(upTimeSeconds), heartBeat.TotalUpTime)
+	assert.Equal(t, int64(downTimeSeconds), heartBeat.TotalDownTime)
 }
